revert cpufreq_interactive patch, undefined error

+++ drivers/cpufreq/cpufreq_interactive.c
@@ -454,8 +454,8 @@ static u64 update_load(int cpu)
 	struct cpufreq_interactive_cpuinfo *pcpu = &per_cpu(cpuinfo, cpu);
 	u64 now;
 	u64 now_idle;
-	u64 delta_idle;
-	u64 delta_time;
+	unsigned int delta_idle;
+	unsigned int delta_time;
 	u64 active_time;
 #if defined(CONFIG_SEC_PM) || defined(CONFIG_MODE_AUTO_CHANGE)
 	unsigned int cur_load = 0;
@@ -465,8 +465,8 @@ static u64 update_load(int cpu)
 #endif
 
 	now_idle = get_cpu_idle_time(cpu, &now);
-	delta_idle = (now_idle - pcpu->time_in_idle);
-	delta_time = (now - pcpu->time_in_idle_timestamp);
+	delta_idle = (unsigned int)(now_idle - pcpu->time_in_idle);
+	delta_time = (unsigned int)(now - pcpu->time_in_idle_timestamp);
 
 	if (delta_time <= delta_idle)
 		active_time = 0;
@@ -615,7 +615,7 @@ static void exit_mode(void)
 static void cpufreq_interactive_timer(unsigned long data)
 {
 	u64 now;
-	u64 delta_time;
+	unsigned int delta_time;
 	u64 cputime_speedadj;
 	int cpu_load;
 	struct cpufreq_interactive_cpuinfo *pcpu =
@@ -639,7 +639,7 @@ static void cpufreq_interactive_timer(unsigned long data)
 
 	spin_lock_irqsave(&pcpu->load_lock, flags);
 	now = update_load(data);
-	delta_time = (now - pcpu->cputime_speedadj_timestamp);
+	delta_time = (unsigned int)(now - pcpu->cputime_speedadj_timestamp);
 	cputime_speedadj = pcpu->cputime_speedadj;
 	spin_unlock_irqrestore(&pcpu->load_lock, flags);
